Total rewrite of -O1 optimizations
A much superior and much simpler algorithm is used: iterate through the expression backwards

This is precisely what you would do if you were simplifying the expression with paper and pen. You would compare the term you're adding with the last term in the expression, make any changes that were necessary, move the term up, and repeat until it was in the desired position.
This is now precisely what we do here -- and as a result, all tests are finally passing

The code needs to be cleaned up a little bit -- we're using raw pointers (we shouldn't be), there's some wacky for-loops where something more straightforward should be used, but otherwise, this seems like the way to go.