-O2 insert handling: don't only shift replaces that *start* after the redundancy
We're handling cases we're:
 - We've previously REMOVED from position X to position Y
 - We're now INSERTING to position X
Under -O2, those two instructions are taken to be redundant, and can be optimized to a single REPLACE instruction

When we do this, we have to shift any instructions that came after that remove -- their position data was based on a map of the file that's no longer valid. Ie, if we then REPLACE something at a position >= X, its position was *previously* taking into account that something before it had been removed -- now that that something isn't removed, we need to right-shift it by the length of the redundancy

Previously, we were fixing REPLACEs in particular by checking if their *start* was >= X. However, it could be that we're starting at (for example) X - 1 and replacing several characters, in which case, it doesn't matter that it doesn't start at X -- a *portion* of it still has to be right-shifted.

So we do that now. We figure which part of the instruction (if any) goes after the redundant portion, and we right-shift it

This also fixes the most recently-added failing test